requirements.txt
pandas
PyPDF2
python-docx
python-pptx
openpyxl
